Fix web checkout activity memory leak
The WebView was causing a memory leak when the activity was destroyed,
as diagnosed by the StrictMode.VMPolicy monitoring for Activity leaks in
the example app. The leak appears to be caused by enabling JavaScript
and is fixed by disabling it.

The leak can alternatively be fixed by removing the WebView from the
view hierarchy and destroying it. However this causes the warning
"Application attempted to call on a destroyed WebView" to be printed to
the debug console which is not a desirable property of a public SDK.