Use new build config variable for library version
The latest version of the Android Gradle plugin (4.1) has removed
BuildConfig.VERSION_NAME for library modules. This was done to avoid
confusion with the version name generated for the application, as
justified here:
https://developer.android.com/studio/releases/gradle-plugin#version_properties_removed_from_buildconfig_class_in_library_projects

To work around this the custom build config variable
`AfterpayLibraryVersion` is used to define the current library version
for use in the version header included in the web checkout flow.