Refactor V3 + CAP
I misunderstood where 'isCashAppPay' is needed. It is not something that needs
to be exposed and set by the merchant. It is internal to the SDK.

V3 + CAP doesn't launch any Afterpay Intent / UI. wWe need to expose
the two network requests normally done in the Afterpay UI
(see CheckoutV3ViewModel) directly to the merchant

* perform checkout request
* perform confirmation request

Furthermore we can save the merchant some trouble by combining
the perform checkout request with the sign token request

So we add two new functions to the API:

* beginCheckoutV3WithCashAppPay
* confirmCheckoutV3WithCashAppPay

In the process of this work I create an overload of
AfterpayCashAppCheckout.performSignPaymentRequest which
removes the unnecessary suspend + callback.